requirements.txt
# Trading Platform — dépendances Python
# Installer : pip install -r requirements.txt

# --- Data & Calcul quantitatif ---
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# --- Validation JSON ---
jsonschema>=4.17.0

# --- API IG Markets ---
requests>=2.31.0
aiohttp>=3.9.0

# --- LLM (Research Agent) ---
anthropic>=0.20.0

# --- Configuration ---
python-dotenv>=1.0.0

# --- Logging structuré ---
structlog>=23.1.0

# --- Tests ---
pytest>=7.4.0
pytest-asyncio>=0.21.0

# --- Données de marché réelles ---
yfinance>=0.2.40

# --- Binance (crypto live + earn) ---
python-binance>=1.0.19

# --- Alpaca Markets (paper trading + live US equities) ---
alpaca-py>=0.20.0

# --- IBKR (FX/EU/futures live via TWS/Gateway) ---
ib_insync>=0.9.86

# --- System utilities (process info, timezones, YAML config) ---
psutil>=5.9.0
pytz>=2023.3
PyYAML>=6.0

# --- Dashboard Pro (interface locale) ---
streamlit>=1.35.0
plotly>=5.18.0

# --- Dashboard API (FastAPI) ---
fastapi>=0.104.0
uvicorn>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# --- Scheduling ---
schedule>=1.2.0

# --- Intraday backtester ---
scikit-learn>=1.3.0
pyarrow>=14.0.0
ta>=0.11.0

# --- Optionnel (backtest avancé) ---
# vectorbt>=0.26.0   # Décommenter pour backtests haute performance
